+++ app/Http/Livewire/VehiculosController.php
@@ -7,6 +7,7 @@ use Livewire\Component;
 use App\Models\Vehiculos;
 use App\Models\Dependencia;
 use App\Models\Conductor;
+use App\Models\Diagnostico;
 use DB;
 
 class VehiculosController extends Component
@@ -77,18 +78,18 @@ class VehiculosController extends Component
     }
     public function Store()
     {
-      /*
+      
         $rules = [
             'placa' => 'required|min:3|unique:vehiculos,placa',
             'clase' => 'required|min:3',
             'marca' => 'required|min:3',
             'tipo_vehiculo' => 'required|min:3',
-            'color' => 'required|min:3',
-            'combustible_capacidad' => 'required|min:3',
-            'motor' => 'required|min:3|unique:vehiculos,motor',
-            'chasis' => 'required|min:3|unique:vehiculos,chasis',
+            'color' => '',
+            'combustible_capacidad' => '',
+            'motor' => '',
+            'chasis' => '',
             'modelo' => 'required|min:3',
-            'cilindrada' => 'required|min:3',
+            'cilindrada' => '',
             'estado' => 'required|min:3',
             'dependencias_id' => 'required',
             //'conductors_id' => 'required',
@@ -101,7 +102,7 @@ class VehiculosController extends Component
             
         ];
 
-        $this->validate($rules, $messages);*/
+        $this->validate($rules, $messages);
 
         $Vehiculos=Vehiculos::create(['placa' => $this->placa,
                                       'clase' => $this->clase,
@@ -145,14 +146,17 @@ class VehiculosController extends Component
     public function Update()
     {
         $rules = [
-            'placa' => "required|min:3|unique:vehiculos,placa,{$this->selected_id}",
-            'modelo' => 'required|min:3',
+            'placa' => 'required|min:3',
+            'clase' => 'required|min:3',
             'marca' => 'required|min:3',
-            'color' => 'required|min:3',
-            'año' => 'required|min:3',
-            'cilindrada' => 'required|min:3',
-            'chasis' => 'required|min:3',
-            'motor' => 'required|min:3',
+            'tipo_vehiculo' => 'required|min:3',
+            'color' => '',
+            'combustible_capacidad' => '',
+            'motor' => '',
+            'chasis' => '',
+            'modelo' => 'required|min:3',
+            'cilindrada' => '',
+            'estado' => 'required|min:3',
             'dependencias_id' => 'required',
             //'conductors_id' => 'required',
         ];
@@ -167,28 +171,38 @@ class VehiculosController extends Component
         $this->validate($rules, $messages);
         $Vehiculos =Vehiculos::find($this->selected_id);
         $Vehiculos ->update(['placa' => $this->placa,
-                                      'modelo'=>$this->modelo,
-                                      'marca'=>$this->marca,
-                                      'color'=>$this->color,
-                                      'año'=>$this->año,
-                                      'cilindrada'=>$this->cilindrada,
-                                      'chasis'=>$this->chasis,
-                                      'motor'=>$this->motor,
-                                      'dependencias_id' => $this->dependencias_id]);
+                                'clase' => $this->clase,
+                                'marca'=>$this->marca,
+                                'tipo_vehiculo'=>$this->tipo_vehiculo,
+                                'color'=>$this->color,
+                                'combustible_capacidad' => $this->combustible_capacidad,
+                                'motor'=>$this->motor,
+                                'chasis'=>$this->chasis,
+                                'modelo'=>$this->modelo,
+                                'cilindrada'=>$this->cilindrada,
+                                'estado'=>$this->estado,
+                                'dependencias_id' => $this->dependencias_id]);
                                       //'conductors_id' => $this->conductors_id]);
                                       
        // dd($Dependencias);
         $this->resetUI();
         $this->emit('vehiculo-updated', 'Se actualizo el vehiculo con exito');
     }
-    protected $listeners=['deleteRow'=>'Destroy'];
+    protected $listeners=['destroy'=>'Destroy'];
 
     public function Destroy(Vehiculos $Vehiculos)
     {
-        $this->selected_id = $Vehiculos->id;
-        $Vehiculos ->delete();
-        $this->resetUI();
-        $this->emit('vehiculo-deleted', 'Se elimino el vehiculo');
+        if($Vehiculos){
+            $data = Diagnostico::where('vehiculos_id', $Vehiculos->id)->count();
+            //validar si tiene vehiculos asignados
+            if($data > 0 ) {
+                $this->emit('vehiculo-deleted','No es posible eliminar el vehiculo por que tiene diagnosticos ');
+            } else {
+                $Vehiculos->delete();
+                $this->resetUI();
+                $this->emit('vehiculo-deleted', 'Vehiculo Eliminado');
+            }
+        }
     
     }
